refactor: Dispatch raw parameters to DSP

+++ app/src/main/java/com/llsl/viper4android/data/model/DsPreset.kt
@@ -19,6 +19,6 @@ data class DsPreset(
     @ColumnInfo(name = "x_high") val xHigh: Int,
     @ColumnInfo(name = "y_low") val yLow: Int,
     @ColumnInfo(name = "y_high") val yHigh: Int,
-    @ColumnInfo(name = "side_gain_low") val sideGainLow: Int,
-    @ColumnInfo(name = "side_gain_high") val sideGainHigh: Int,
+    @ColumnInfo(name = "side_gain_low") val sideGainLow: Float,
+    @ColumnInfo(name = "side_gain_high") val sideGainHigh: Float,
 )
